+++ src/WebJobs.Script/ScriptHostBuilderExtensions.cs
@@ -3,7 +3,11 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Diagnostics.Tracing;
 using System.Runtime.InteropServices;
+using System.Threading;
+using Azure.Identity;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Channel;
 using Microsoft.ApplicationInsights.DataContracts;
@@ -12,7 +16,6 @@ using Microsoft.ApplicationInsights.Extensibility.Implementation;
 using Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel;
 using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Azure.WebJobs.Host.Executors;
-using Microsoft.Azure.WebJobs.Host.Scale;
 using Microsoft.Azure.WebJobs.Hosting;
 using Microsoft.Azure.WebJobs.Logging;
 using Microsoft.Azure.WebJobs.Logging.ApplicationInsights;
@@ -42,6 +45,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Options;
+using TokenCredentialOptions = Microsoft.Azure.WebJobs.Logging.ApplicationInsights.TokenCredentialOptions;
 
 namespace Microsoft.Azure.WebJobs.Script
 {
@@ -232,8 +236,12 @@ namespace Microsoft.Azure.WebJobs.Script
                     }
                     catch (Exception ex)
                     {
-                        string appInsightsConnStr = context.Configuration[EnvironmentSettingNames.AppInsightsConnectionString];
-                        RecordAndThrowExternalStartupException("Error configuring services in an external startup class.", ex, loggerFactory, appInsightsConnStr);
+                        string appInsightsConnStr = GetConfigurationValue(EnvironmentSettingNames.AppInsightsConnectionString, context.Configuration);
+                        if (!string.IsNullOrEmpty(appInsightsConnStr))
+                        {
+                            string appInsightsAuthStr = GetConfigurationValue(EnvironmentSettingNames.AppInsightsAuthenticationString, context.Configuration);
+                            RecordAndThrowExternalStartupException("Error configuring services in an external startup class.", ex, loggerFactory, appInsightsConnStr, appInsightsAuthStr);
+                        }
                     }
                 }
 
@@ -260,8 +268,12 @@ namespace Microsoft.Azure.WebJobs.Script
                         catch (Exception ex)
                         {
                             // Go directly to the environment; We have no valid configuration from the customer at this point.
-                            string appInsightsConnStr = Environment.GetEnvironmentVariable(EnvironmentSettingNames.AppInsightsConnectionString);
-                            RecordAndThrowExternalStartupException("Error building configuration in an external startup class.", ex, loggerFactory, appInsightsConnStr);
+                            string appInsightsConnStr = GetConfigurationValue(EnvironmentSettingNames.AppInsightsConnectionString);
+                            if (!string.IsNullOrEmpty(appInsightsConnStr))
+                            {
+                                string appInsightsAuthStr = GetConfigurationValue(EnvironmentSettingNames.AppInsightsAuthenticationString);
+                                RecordAndThrowExternalStartupException("Error building configuration in an external startup class.", ex, loggerFactory, appInsightsConnStr, appInsightsAuthStr);
+                            }
                         }
                     });
                 }
@@ -402,16 +414,37 @@ namespace Microsoft.Azure.WebJobs.Script
 
         internal static void ConfigureApplicationInsights(HostBuilderContext context, ILoggingBuilder builder)
         {
-            string appInsightsInstrumentationKey = context.Configuration[EnvironmentSettingNames.AppInsightsInstrumentationKey];
-            string appInsightsConnectionString = context.Configuration[EnvironmentSettingNames.AppInsightsConnectionString];
+            string appInsightsInstrumentationKey = GetConfigurationValue(EnvironmentSettingNames.AppInsightsInstrumentationKey, context.Configuration);
+            string appInsightsConnectionString = GetConfigurationValue(EnvironmentSettingNames.AppInsightsConnectionString, context.Configuration);
 
             // Initializing AppInsights services during placeholder mode as well to avoid the cost of JITting these objects during specialization
             if (!string.IsNullOrEmpty(appInsightsInstrumentationKey) || !string.IsNullOrEmpty(appInsightsConnectionString) || SystemEnvironment.Instance.IsPlaceholderModeEnabled())
             {
+                string eventLogLevel = GetConfigurationValue(EnvironmentSettingNames.AppInsightsEventListenerLogLevel, context.Configuration);
+                string authString = GetConfigurationValue(EnvironmentSettingNames.AppInsightsAuthenticationString, context.Configuration);
+                // Initialize ScriptTelemetryInitializer before any other telemetry initializers.
+                // This will allow HostInstanceId to be removed as part of MetricsCustomDimensionOptimization.
+                builder.Services.AddSingleton<ITelemetryInitializer, ScriptTelemetryInitializer>();
                 builder.AddApplicationInsightsWebJobs(o =>
                 {
                     o.InstrumentationKey = appInsightsInstrumentationKey;
                     o.ConnectionString = appInsightsConnectionString;
+
+                    if (!string.IsNullOrEmpty(eventLogLevel))
+                    {
+                        if (Enum.TryParse(eventLogLevel, ignoreCase: true, out EventLevel level))
+                        {
+                            o.DiagnosticsEventListenerLogLevel = level;
+                        }
+                        else
+                        {
+                            throw new InvalidEnumArgumentException($"Invalid `{EnvironmentSettingNames.AppInsightsEventListenerLogLevel}`.");
+                        }
+                    }
+                    if (!string.IsNullOrEmpty(authString))
+                    {
+                        o.TokenCredentialOptions = TokenCredentialOptions.ParseAuthenticationString(authString);
+                    }
                 }, t =>
                 {
                     if (t.TelemetryChannel is ServerTelemetryChannel channel)
@@ -425,7 +458,6 @@ namespace Microsoft.Azure.WebJobs.Script
 
                 builder.Services.ConfigureOptions<ApplicationInsightsLoggerOptionsSetup>();
                 builder.Services.AddSingleton<ISdkVersionProvider, FunctionsSdkVersionProvider>();
-                builder.Services.AddSingleton<ITelemetryInitializer, ScriptTelemetryInitializer>();
 
                 if (SystemEnvironment.Instance.IsPlaceholderModeEnabled())
                 {
@@ -502,7 +534,7 @@ namespace Microsoft.Azure.WebJobs.Script
             }
         }
 
-        private static void RecordAndThrowExternalStartupException(string message, Exception ex, ILoggerFactory loggerFactory, string appInsightsConnStr)
+        private static void RecordAndThrowExternalStartupException(string message, Exception ex, ILoggerFactory loggerFactory, string appInsightsConnStr, string appInsightsAuthString)
         {
             var startupEx = new ExternalStartupException(message, ex);
 
@@ -511,22 +543,44 @@ namespace Microsoft.Azure.WebJobs.Script
 
             // Send the error to App Insights if possible. This is happening during ScriptHost construction so we
             // have no existing TelemetryClient to use. Create a one-off client and flush it ASAP.
-            if (appInsightsConnStr is not null)
+            using TelemetryConfiguration telemetryConfiguration = new()
             {
-                using TelemetryConfiguration telemetryConfiguration = new()
-                {
-                    ConnectionString = appInsightsConnStr,
-                    TelemetryChannel = new InMemoryChannel()
-                };
+                ConnectionString = appInsightsConnStr,
+                TelemetryChannel = new InMemoryChannel()
+            };
 
-                TelemetryClient telemetryClient = new(telemetryConfiguration);
-                telemetryClient.Context.GetInternalContext().SdkVersion = new ApplicationInsightsSdkVersionProvider().GetSdkVersion();
-                telemetryClient.TrackTrace(startupEx.ToString(), SeverityLevel.Error);
-                telemetryClient.TrackException(startupEx);
-                telemetryClient.Flush();
+            if (!string.IsNullOrEmpty(appInsightsAuthString))
+            {
+                TokenCredentialOptions credentialOptions = TokenCredentialOptions.ParseAuthenticationString(appInsightsAuthString);
+                // Default is connection string based ingestion
+                telemetryConfiguration.SetAzureTokenCredential(new ManagedIdentityCredential(credentialOptions.ClientId));
             }
 
+            TelemetryClient telemetryClient = new(telemetryConfiguration);
+            telemetryClient.Context.GetInternalContext().SdkVersion = new ApplicationInsightsSdkVersionProvider().GetSdkVersion();
+            telemetryClient.TrackTrace(startupEx.ToString(), SeverityLevel.Error);
+            telemetryClient.TrackException(startupEx);
+            telemetryClient.Flush();
+            // Flush is async, sleep for 1 sec to give it time to flush
+            Thread.Sleep(1000);
+
             throw startupEx;
+        }
+
+        private static string GetConfigurationValue(string key, IConfiguration configuration = null)
+        {
+            if (configuration != null && configuration[key] is string configValue)
+            {
+                return configValue;
+            }
+            else if (Environment.GetEnvironmentVariable(key) is string envValue)
+            {
+                return envValue;
+            }
+            else
+            {
+                return null;
+            }
         }
     }
 }